microservices/salle-service/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
kafka-python==2.0.2
